Fold the moderation no-op into the contract; drop default.ts
The default.ts / plugin.ts re-export split was the price of giving the
enterprise gate an OSS-reachable fallback. Collapse it: the no-op is just the
contract's empty value, so it lives in types.ts as PLUGIN_MODERATION_NOOP. The
OSS plugin.ts re-exports it, and the enterprise gate delegates to it. OSS is
back to two files (types + a one-line plugin entry), and nothing in frontend/src
is enterprise-aware.
