+++ frontend/src/metabase/moderation/plugin.ts
@@ -3,8 +3,8 @@
  *
  *   import { PLUGIN_MODERATION } from "metabase/moderation/plugin";
  *
- * In OSS builds this re-exports the default (no-op) implementation. In
- * enterprise builds the whole module is swapped for
- * `metabase-enterprise/moderation/plugin` (see resolve-aliases.js).
+ * In OSS builds this is the contract's no-op value. In enterprise builds the
+ * whole module is swapped for `metabase-enterprise/moderation/plugin` (see
+ * resolve-aliases.js).
  */
-export { PLUGIN_MODERATION } from "./default";
+export { PLUGIN_MODERATION_NOOP as PLUGIN_MODERATION } from "./types";
